+++ release_notes.rst
@@ -6,8 +6,11 @@ Bugfixes:
     - remove closing of stdout filehandle as it might lead to stacktraces when
       its still required. Leave it as is.
 
+    - Fixed wishbone.protocol.decoder problems when exceeding allowed buffer.
+
 Features:
 
     - Changed behaviour of Actor.setEncoder and Actor.setDecoder so they don't
       override the referred encoder/decoder when one is already setup through
       ActorConfig.
+